refactor(batch): move batch subsystem out of _core into the buttermilk/batch addon; drop misleading "vertex" module name
KEEP all batch functionality (multi-provider Vertex Gemini/Claude/Llama/DeepSeek
via google.genai SDK + OpenAI/Azure batch) but relocate it so it is clearly an
OPTIONAL ADDON, decoupled from the default (online) call path. Pure structural
move/rename + import rewiring — no behaviour change.

Moves:
- buttermilk/_core/vertex_batch.py   -> buttermilk/batch/managers.py
- buttermilk/processors/vertex_batch.py -> buttermilk/batch/processors.py

The "vertex" module name was a misnomer (both modules handle OpenAI/Azure too).
New names are provider-neutral and match the existing flat-module convention in
buttermilk/batch/ (result.py, runner.py, executors/). Public class names are
unchanged (VertexBatchProcessor genuinely is the Vertex subclass), so
buttermilk.processors.{VertexBatchProcessor,OpenAIBatchProcessor,BatchLLMProcessor}
keep working via re-export.

Dependency direction is now ADDON -> CORE only:
- buttermilk/_core no longer hosts any batch module.
- _core/llms.py and the standard online agent path do not import the addon.
- The addon keeps importing from _core (to_litellm_messages, LLMConfig,
  litellm_provider_segment, save/pricing utils) — the correct direction.

Updated all importers repo-wide (src + cli + scripts + tests) and relocated the
batch unit tests under tests/batch/ to mirror the package.
