+++ buttermilk/batch/executors/vertex.py
@@ -2,8 +2,8 @@ from buttermilk import logger
 from buttermilk._core.processing_context import ProcessingContext
 from buttermilk._core.processor_core import BatchProcessorCore
 from buttermilk._core.types import BaseRecord
-from buttermilk._core.vertex_batch import BatchJobManager
 from buttermilk.batch.executors.base import BatchExecutor
+from buttermilk.batch.managers import BatchJobManager
 from buttermilk.batch.result import BatchExecutionResult, BatchJobStatus
