+++ buttermilk/batch/processors.py
@@ -47,14 +47,14 @@ from buttermilk._core.exceptions import FatalError
 from buttermilk._core.processing_context import ProcessingContext
 from buttermilk._core.processor_core import BatchProcessorCore, TraceParams
 from buttermilk._core.types import BaseRecord
-from buttermilk._core.vertex_batch import BatchJobManager, BatchResult, OpenAIBatchJobManager
+from buttermilk.batch.managers import BatchJobManager, BatchResult, OpenAIBatchJobManager
 from buttermilk.utils.import_utils import load_class
 from buttermilk.utils.templating import make_messages, render_template
 
 if TYPE_CHECKING:
     from google.genai.types import BatchJob
 
-    from buttermilk._core.vertex_batch import BatchJobManager
+    from buttermilk.batch.managers import BatchJobManager
 
 
 from buttermilk._core.llms import litellm_provider_segment as _provider_segment  # dedup: single source in llms.py
@@ -405,7 +405,7 @@ class BatchLLMProcessor(BatchProcessorCore):
             List of BatchRequest objects
         """
         from buttermilk._core.llms import to_litellm_messages
-        from buttermilk._core.vertex_batch import BatchRequest
+        from buttermilk.batch.managers import BatchRequest
 
         requests: list[BatchRequest] = []
 
@@ -713,7 +713,7 @@ class VertexBatchProcessor(BatchLLMProcessor):
             return None
 
         from buttermilk._core.json_schema import prepare_schema_for_vertex
-        from buttermilk._core.vertex_batch import _is_claude_model, _is_llama_model
+        from buttermilk.batch.managers import _is_claude_model, _is_llama_model
 
         # Gemini requires enum values converted to strings; Claude and Llama do not
         is_gemini = not _is_claude_model(self.model) and not _is_llama_model(self.model)
@@ -743,7 +743,7 @@ class VertexBatchProcessor(BatchLLMProcessor):
 
     def _ensure_manager(self) -> BatchJobManager:
         """Lazily initialize the BatchJobManager via buttermilk infrastructure."""
-        from buttermilk._core.vertex_batch import BatchJobManager as _BatchJobManager
+        from buttermilk.batch.managers import BatchJobManager as _BatchJobManager
 
         if self._manager is None:
             from google import genai
@@ -1110,7 +1110,7 @@ class OpenAIBatchProcessor(BatchLLMProcessor):
             return self._manager
 
         from buttermilk import bm
-        from buttermilk._core.vertex_batch import OpenAIBatchJobManager
+        from buttermilk.batch.managers import OpenAIBatchJobManager
 
         client = self._ensure_openai_client()
         config = bm.llms.connections[self.model]
