+++ tests/batch/test_decoupled_workflow.py
@@ -17,7 +17,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from buttermilk._core.vertex_batch import (
+from buttermilk.batch.managers import (
     BatchJobManager,
     BatchJobManifest,
     BatchRequest,
