+++ tests/batch/test_executor_vertex.py
