+++ tests/batch/test_managers_openai.py
@@ -10,7 +10,7 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from buttermilk._core.vertex_batch import (
+from buttermilk.batch.managers import (
     BatchRequest,
     OpenAIBatchJobManager,
     OpenAIBatchManifest,
